+++ py-polars/tests/unit/datatypes/test_decimal.py
@@ -6,9 +6,7 @@ from dataclasses import dataclass
 from decimal import Decimal as D
 from typing import Any, NamedTuple
 
-import numpy as np
 import pytest
-from numpy.testing import assert_array_equal
 
 import polars as pl
 from polars.testing import assert_frame_equal
@@ -298,20 +296,3 @@ def test_decimal_write_parquet_12375() -> None:
     assert df["bye"].dtype == pl.Decimal
 
     df.write_parquet(f)
-
-
-@pytest.mark.parametrize("use_pyarrow", [True, False])
-def test_decimal_numpy_export(use_pyarrow: bool) -> None:
-    decimal_data = [D("1.234"), D("2.345"), D("-3.456")]
-
-    s = pl.Series("n", decimal_data)
-    df = s.to_frame()
-
-    assert_array_equal(
-        np.array(decimal_data),
-        s.to_numpy(use_pyarrow=use_pyarrow),
-    )
-    assert_array_equal(
-        np.array(decimal_data).reshape((-1, 1)),
-        df.to_numpy(use_pyarrow=use_pyarrow),
-    )
