+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -287,39 +287,6 @@ def test_int_to_python_timedelta() -> None:
     ).rows() == [(100001, 100001, 100001), (200002, 200002, 200002)]
 
 
-def test_from_numpy() -> None:
-    # note: numpy timeunit support is limited to those supported by polars.
-    # as a result, datetime64[s] raises
-    x = np.asarray(range(100_000, 200_000, 10_000), dtype="datetime64[s]")
-    with pytest.raises(ValueError, match="Please cast to the closest supported unit"):
-        pl.Series(x)
-
-
-@pytest.mark.parametrize(
-    ("numpy_time_unit", "expected_values", "expected_dtype"),
-    [
-        ("ns", ["1970-01-02T01:12:34.123456789"], pl.Datetime("ns")),
-        ("us", ["1970-01-02T01:12:34.123456"], pl.Datetime("us")),
-        ("ms", ["1970-01-02T01:12:34.123"], pl.Datetime("ms")),
-        ("D", ["1970-01-02"], pl.Date),
-    ],
-)
-def test_from_numpy_supported_units(
-    numpy_time_unit: str,
-    expected_values: list[str],
-    expected_dtype: PolarsTemporalType,
-) -> None:
-    values = np.array(
-        ["1970-01-02T01:12:34.123456789123456789"],
-        dtype=f"datetime64[{numpy_time_unit}]",
-    )
-    result = pl.from_numpy(values)
-    expected = (
-        pl.Series("column_0", expected_values).str.strptime(expected_dtype).to_frame()
-    )
-    assert_frame_equal(result, expected)
-
-
 def test_datetime_consistency() -> None:
     dt = datetime(2022, 7, 5, 10, 30, 45, 123455)
     df = pl.DataFrame({"date": [dt]})
@@ -478,36 +445,6 @@ def test_rows() -> None:
     rows = df.rows()
     assert rows[0][0] == date(2308, 4, 2)
     assert rows[0][1] == datetime(1970, 1, 1, 0, 2, 3, 543000)
-
-
-def test_series_to_numpy() -> None:
-    s0 = pl.Series("date", [123543, 283478, 1243]).cast(pl.Date)
-    s1 = pl.Series(
-        "datetime", [datetime(2021, 1, 2, 3, 4, 5), datetime(2021, 2, 3, 4, 5, 6)]
-    )
-    s2 = pl.datetime_range(
-        datetime(2021, 1, 1, 0),
-        datetime(2021, 1, 1, 1),
-        interval="1h",
-        time_unit="ms",
-        eager=True,
-    )
-    assert str(s0.to_numpy()) == "['2308-04-02' '2746-02-20' '1973-05-28']"
-    assert (
-        str(s1.to_numpy()[:2])
-        == "['2021-01-02T03:04:05.000000' '2021-02-03T04:05:06.000000']"
-    )
-    assert (
-        str(s2.to_numpy()[:2])
-        == "['2021-01-01T00:00:00.000' '2021-01-01T01:00:00.000']"
-    )
-    s3 = pl.Series([timedelta(hours=1), timedelta(hours=-2)])
-    out = np.array([3_600_000_000_000, -7_200_000_000_000], dtype="timedelta64[ns]")
-    assert (s3.to_numpy() == out).all()
-
-    s4 = pl.Series([time(10, 30, 45), time(23, 59, 59)])
-    out = np.array([time(10, 30, 45), time(23, 59, 59)], dtype="object")
-    assert (s4.to_numpy() == out).all()
 
 
 @pytest.mark.parametrize(
