+++ py-polars/tests/unit/interop/test_interop.py
@@ -7,7 +7,6 @@ import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
-from numpy.testing import assert_array_equal
 
 import polars as pl
 from polars.exceptions import ComputeError
@@ -15,79 +14,6 @@ from polars.testing import assert_frame_equal, assert_series_equal
 
 if TYPE_CHECKING:
     from polars.type_aliases import PolarsDataType
-
-
-@pytest.fixture(
-    params=[
-        ("int8", [1, 3, 2], pl.Int8, np.int8),
-        ("int16", [1, 3, 2], pl.Int16, np.int16),
-        ("int32", [1, 3, 2], pl.Int32, np.int32),
-        ("int64", [1, 3, 2], pl.Int64, np.int64),
-        ("uint8", [1, 3, 2], pl.UInt8, np.uint8),
-        ("uint16", [1, 3, 2], pl.UInt16, np.uint16),
-        ("uint32", [1, 3, 2], pl.UInt32, np.uint32),
-        ("uint64", [1, 3, 2], pl.UInt64, np.uint64),
-        ("float32", [21.7, 21.8, 21], pl.Float32, np.float32),
-        ("float64", [21.7, 21.8, 21], pl.Float64, np.float64),
-        ("bool", [True, False, False], pl.Boolean, np.bool_),
-        ("object", [21.7, "string1", object()], pl.Object, np.object_),
-        ("str", ["string1", "string2", "string3"], pl.String, np.str_),
-        ("intc", [1, 3, 2], pl.Int32, np.intc),
-        ("uintc", [1, 3, 2], pl.UInt32, np.uintc),
-        ("str_fixed", ["string1", "string2", "string3"], pl.String, np.str_),
-        (
-            "bytes",
-            [b"byte_string1", b"byte_string2", b"byte_string3"],
-            pl.Binary,
-            np.bytes_,
-        ),
-    ]
-)
-def numpy_interop_test_data(request: Any) -> Any:
-    return request.param
-
-
-def test_df_from_numpy(numpy_interop_test_data: Any) -> None:
-    name, values, pl_dtype, np_dtype = numpy_interop_test_data
-    df = pl.DataFrame({name: np.array(values, dtype=np_dtype)})
-    assert [pl_dtype] == df.dtypes
-
-
-def test_asarray(numpy_interop_test_data: Any) -> None:
-    name, values, pl_dtype, np_dtype = numpy_interop_test_data
-    pl_series_to_numpy_array = np.asarray(pl.Series(name, values, pl_dtype))
-    numpy_array = np.asarray(values, dtype=np_dtype)
-    assert_array_equal(pl_series_to_numpy_array, numpy_array)
-
-
-@pytest.mark.parametrize("use_pyarrow", [True, False])
-def test_to_numpy(numpy_interop_test_data: Any, use_pyarrow: bool) -> None:
-    name, values, pl_dtype, np_dtype = numpy_interop_test_data
-    pl_series_to_numpy_array = pl.Series(name, values, pl_dtype).to_numpy(
-        use_pyarrow=use_pyarrow
-    )
-    numpy_array = np.asarray(values, dtype=np_dtype)
-    assert_array_equal(pl_series_to_numpy_array, numpy_array)
-
-
-@pytest.mark.parametrize("use_pyarrow", [True, False])
-@pytest.mark.parametrize("has_null", [True, False])
-@pytest.mark.parametrize("dtype", [pl.Time, pl.Boolean, pl.String])
-def test_to_numpy_no_zero_copy(
-    use_pyarrow: bool, has_null: bool, dtype: pl.PolarsDataType
-) -> None:
-    data: list[Any] = ["a", None] if dtype == pl.String else [0, None]
-    series = pl.Series(data if has_null else data[:1], dtype=dtype)
-    with pytest.raises(ValueError):
-        series.to_numpy(zero_copy_only=True, use_pyarrow=use_pyarrow)
-
-
-def test_to_numpy_empty_no_pyarrow() -> None:
-    series = pl.Series([], dtype=pl.Null)
-    result = series.to_numpy()
-    assert result.dtype == pl.Float32
-    assert result.shape == (0,)
-    assert result.size == 0
 
 
 def test_from_pandas() -> None:
@@ -449,109 +375,6 @@ def test_from_records() -> None:
     assert df.rows() == [(1, 4), (2, 5), (3, 6)]
 
 
-def test_from_numpy() -> None:
-    data = np.array([[1, 2, 3], [4, 5, 6]])
-    df = pl.from_numpy(
-        data,
-        schema=["a", "b"],
-        orient="col",
-        schema_overrides={"a": pl.UInt32, "b": pl.UInt32},
-    )
-    assert df.shape == (3, 2)
-    assert df.rows() == [(1, 4), (2, 5), (3, 6)]
-    assert df.schema == {"a": pl.UInt32, "b": pl.UInt32}
-    data2 = np.array(["foo", "bar"], dtype=object)
-    df2 = pl.from_numpy(data2)
-    assert df2.shape == (2, 1)
-    assert df2.rows() == [("foo",), ("bar",)]
-    assert df2.schema == {"column_0": pl.String}
-    with pytest.raises(
-        ValueError,
-        match="cannot create DataFrame from array with more than two dimensions",
-    ):
-        _ = pl.from_numpy(np.array([[[1]]]))
-    with pytest.raises(
-        ValueError, match="cannot create DataFrame from zero-dimensional array"
-    ):
-        _ = pl.from_numpy(np.array(1))
-
-
-def test_from_numpy_array_value() -> None:
-    df = pl.DataFrame({"A": [[2, 3]]})
-    assert df.rows() == [([2, 3],)]
-    assert df.schema == {"A": pl.List(pl.Int64)}
-
-
-def test_construct_from_nparray_value() -> None:
-    array_cell = np.array([2, 3])
-    df = pl.DataFrame(np.array([[array_cell, 4]], dtype=object))
-    assert df.dtypes == [pl.Object, pl.Object]
-    to_numpy = df.to_numpy()
-    assert to_numpy.shape == (1, 2)
-    assert_array_equal(to_numpy[0][0], array_cell)
-    assert to_numpy[0][1] == 4
-
-
-def test_from_numpy_nparray_value() -> None:
-    array_cell = np.array([2, 3])
-    df = pl.from_numpy(np.array([[array_cell, 4]], dtype=object))
-    assert df.dtypes == [pl.Object, pl.Object]
-    to_numpy = df.to_numpy()
-    assert to_numpy.shape == (1, 2)
-    assert_array_equal(to_numpy[0][0], array_cell)
-    assert to_numpy[0][1] == 4
-
-
-def test_from_numpy_structured() -> None:
-    test_data = [
-        ("Google Pixel 7", 521.90, True),
-        ("Apple iPhone 14 Pro", 999.00, True),
-        ("Samsung Galaxy S23 Ultra", 1199.99, False),
-        ("OnePlus 11", 699.00, True),
-    ]
-    # create a numpy structured array...
-    arr_structured = np.array(
-        test_data,
-        dtype=np.dtype(
-            [
-                ("product", "U32"),
-                ("price_usd", "float64"),
-                ("in_stock", "bool"),
-            ]
-        ),
-    )
-    # ...and also establish as a record array view
-    arr_records = arr_structured.view(np.recarray)
-
-    # confirm that we can cleanly initialise a DataFrame from both,
-    # respecting the native dtypes and any schema overrides, etc.
-    for arr in (arr_structured, arr_records):
-        df = pl.DataFrame(data=arr).sort(by="price_usd", descending=True)
-
-        assert df.schema == {
-            "product": pl.String,
-            "price_usd": pl.Float64,
-            "in_stock": pl.Boolean,
-        }
-        assert df.rows() == sorted(test_data, key=lambda row: -row[1])
-
-        for df in (
-            pl.DataFrame(
-                data=arr, schema=["phone", ("price_usd", pl.Float32), "available"]
-            ),
-            pl.DataFrame(
-                data=arr,
-                schema=["phone", "price_usd", "available"],
-                schema_overrides={"price_usd": pl.Float32},
-            ),
-        ):
-            assert df.schema == {
-                "phone": pl.String,
-                "price_usd": pl.Float32,
-                "available": pl.Boolean,
-            }
-
-
 def test_from_arrow() -> None:
     data = pa.table({"a": [1, 2, 3], "b": [4, 5, 6]})
     df = pl.from_arrow(data)
@@ -632,13 +455,6 @@ def test_no_rechunk() -> None:
     assert pl.from_arrow(table, rechunk=False).n_chunks() == 2
     # chunked array
     assert pl.from_arrow(table["x"], rechunk=False).n_chunks() == 2
-
-
-def test_numpy_to_lit() -> None:
-    out = pl.select(pl.lit(np.array([1, 2, 3]))).to_series().to_list()
-    assert out == [1, 2, 3]
-    out = pl.select(pl.lit(np.float32(0))).to_series().to_list()
-    assert out == [0.0]
 
 
 def test_from_empty_pandas() -> None:
@@ -730,12 +546,6 @@ def test_from_pyarrow_chunked_array() -> None:
     assert series.to_list() == [1, 2]
 
 
-def test_numpy_preserve_uint64_4112() -> None:
-    df = pl.DataFrame({"a": [1, 2, 3]}).with_columns(pl.col("a").hash())
-    assert df.to_numpy().dtype == np.dtype("uint64")
-    assert df.to_numpy(structured=True).dtype == np.dtype([("a", "uint64")])
-
-
 def test_arrow_list_null_5697() -> None:
     # Create a pyarrow table with a list[null] column.
     pa_table = pa.table([[[None]]], names=["mycol"])
@@ -776,29 +586,6 @@ def test_from_pyarrow_map() -> None:
             [{"key": "a", "value": "else"}, {"key": "b", "value": "another key"}],
         ],
     }
-
-
-def test_to_numpy_datelike() -> None:
-    s = pl.Series(
-        "dt",
-        [
-            datetime(2022, 7, 5, 10, 30, 45, 123456),
-            None,
-            datetime(2023, 2, 5, 15, 22, 30, 987654),
-        ],
-    )
-    assert str(s.to_numpy()) == str(
-        np.array(
-            ["2022-07-05T10:30:45.123456", "NaT", "2023-02-05T15:22:30.987654"],
-            dtype="datetime64[us]",
-        )
-    )
-    assert str(s.drop_nulls().to_numpy()) == str(
-        np.array(
-            ["2022-07-05T10:30:45.123456", "2023-02-05T15:22:30.987654"],
-            dtype="datetime64[us]",
-        )
-    )
 
 
 def test_from_fixed_size_binary_list() -> None:
